#16 Refactor and move files into packages
