+++ app/src/main/java/ca/kendallroth/mileageapp/activities/HomeActivity.java
@@ -1,4 +1,4 @@
-package ca.kendallroth.mileageapp;
+package ca.kendallroth.mileageapp.activities;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -6,6 +6,8 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import ca.kendallroth.mileageapp.R;
 
 /**
  * Main activity after authentication has occurred (automatic or Login activity)
