+++ app/src/main/java/ca/kendallroth/mileageapp/fragments/RegisterFragment.java
@@ -1,4 +1,4 @@
-package ca.kendallroth.mileageapp;
+package ca.kendallroth.mileageapp.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,6 +6,8 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import ca.kendallroth.mileageapp.R;
 
 /**
  * Fragment for enabling a user to register for the app
