+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ca.kendallroth.mileageapp.HomeActivity">
+    tools:context="ca.kendallroth.mileageapp.activities.HomeActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
