+++ app/src/main/res/layout/content_home.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="ca.kendallroth.mileageapp.HomeActivity"
+    tools:context="ca.kendallroth.mileageapp.activities.HomeActivity"
     tools:showIn="@layout/activity_home">
 
 </android.support.constraint.ConstraintLayout>
